feat:finishing edit review front-end
