+++ client/src/Components/Pages/Admin/Reviews/AllReviews.js
@@ -133,6 +133,7 @@ export default class AllReviews extends Component {
   }
 
   render() {
+    console.log("data", this.state.data);
     return (
       <>
         <Table
