+++ client/src/Components/Pages/Review/index.js
@@ -1,11 +1,10 @@
 import React, { Component } from "react";
-import { Formik, Form, Field, ErrorMessage } from "formik";
 import axios from "axios";
+import moment from "moment";
 import { Checkbox, message, Spin, Icon } from "antd";
-import Loading from "./../../Common/AntdComponents/Loading";
+import Loading from "../../Common/AntdComponents/Loading";
 
 import {
-  ReviewWrapper,
   SubmitButton,
   UserAgreement,
   CheckboxWrapper,
@@ -17,13 +16,14 @@ import {
   FormWrapper,
   Level2Header,
   AgreementLabel,
-  LinkSpan
+  LinkSpan,
+  ReviewWrapper
 } from "./Review.style";
 
 import { StyledErrorMessage } from "./Question/Question.style";
 
 import Question from "./Question/index";
-import { organizations } from "./../../../theme";
+import { organizations } from "../../../theme";
 
 import { initQueestionsValues } from "./initialQuestionsValues";
 import { validationSchema } from "./validationSchema";
@@ -39,25 +39,19 @@ const antIcon = (
   <Icon type="loading" style={{ fontSize: 24, color: "white" }} spin />
 );
 
-const {
-  API_GET_QUESTIONS_URL,
-  API_POST_REVIEW_URL
-} = require("../../../apiUrls");
-
-// For rate question to add the Org. category
-let rateQ = {};
+const { API_POST_REVIEW_URL } = require("../../../apiUrls");
 
 class Review extends Component {
   state = {
     isLoading: true,
     groups: [],
     groupss: {},
-    organization: { category: "", name: "", needsVerification: false },
+    organization: { category: "agency", name: "", orgId: "" },
+    reviewId: "",
     user: { email: "" },
-    worksiteImage: "",
     dropdownList: [],
-    questions: {},
     comments: {},
+    answers: {},
     review: {
       workPeriod: {
         from: "",
@@ -68,73 +62,226 @@ class Review extends Component {
       // voiceReview: ""
     },
     hasAgreed: false,
-    answers: {}
-  };
-
-  handleChange = e => {
-    this.setState({
-      answers: { ...this.state.answers, [e.target.name]: e.target.value }
-    });
+    questions: [],
+    errors: {},
+    isSubmitting: false,
+    isEditing: false,
+    orgId: ""
   };
 
   componentDidMount() {
+    const { orgId, reviewId } = this.props.match.params;
+    if (!orgId && !reviewId) {
+      this.props.history.push("/search");
+    }
     const { email } = this.props;
-    const { category, name, needsVerification } = this.props.location.state;
-    const { organization, user } = this.state;
 
-    // update the static questions to reflect the orgType
-    rateQ = {
-      number: 19,
-      text: "How would you rate this",
-      type: "rate",
-      options: ["Bad", "Poor", "Average", "Great", "Excellent"]
-    };
-    const text = rateQ.text;
-    const newText = `${text} ${category}?`;
-    rateQ.text = newText;
+    this.setState({
+      organization: {
+        orgId
+      },
+      user: { email },
+      reviewId
+    });
 
-    organization.category = category;
-    organization.name = name;
-    organization.needsVerification = needsVerification || false;
-    user.email = email;
+    if (reviewId) {
+      axios
+        .get(`/api/review/${reviewId}/is-edatable`)
+        .then(res => {
+          const { orgId } = res.data;
+          axios
+            .get(`/api/questions/${orgId}`)
+            .then(res => {
+              const answers = {};
+              const { getReviewAnswers: reviewDetails } = res.data;
+              const review = {
+                workPeriod: {
+                  from: moment(reviewDetails[0].workPeriod.from),
+                  to: moment(reviewDetails[0].workPeriod.to)
+                },
+                rate: reviewDetails[0].rate,
+                overallReview: reviewDetails[0].overallReview.text
+                // voiceReview: ""
+              };
 
-    axios
-      .get(API_GET_QUESTIONS_URL, {
-        params: {
-          organization: category
-        }
-      })
-      .then(res => {
-        const groupss = {};
-        res.data.groups.forEach(group => {
-          groupss[group._id] = {
-            title: group.group.text,
-            main: group.questions.filter(question => !question.isDependent),
-            dependant: group.questions.filter(question => question.isDependent)
-          };
+              reviewDetails[0].answers.map(answer => {
+                const {
+                  answer: ans,
+                  question: [question]
+                } = answer;
+                const number = question.number;
+                if (answers[number]) {
+                  // think about this again;
+                  answers[number] = ans;
+                } else {
+                  answers[number] = ans;
+                }
+              });
+              const groupss = {};
+              res.data.groups.forEach(group => {
+                groupss[group._id] = {
+                  title: group.group.text,
+                  main: group.questions.filter(
+                    question => !question.isDependent
+                  ),
+                  dependant: group.questions.filter(
+                    question => question.isDependent
+                  )
+                };
+              });
+              this.setState({
+                isEditing: true,
+                groups: res.data,
+                groupss,
+                isLoading: false,
+                organization: res.data.organization,
+                email,
+                answers,
+                review,
+                dropdownOptions:
+                  res.data.dropDownListData &&
+                  res.data.dropDownListData[0].category
+              });
+            })
+            .catch(err => {
+              // server error 500
+              const error =
+                err.response && err.response.data && err.response.data.error;
+              message.error(error || "Something went wrong");
+            });
+        })
+        .catch(err => {
+          if (err.response.status === 401) {
+            message.error("You can't edit this review");
+            // redirect back to the edit page
+            this.props.history.push("/search");
+          } else {
+            const error =
+              err.response && err.response.data && err.response.data.error;
+            message.error(error || "Something went wrong");
+          }
         });
-        this.setState({
-          groups: res.data,
-          groupss,
-          isLoading: false,
-          organization,
-          user,
-          email,
-          dropdownOptions:
-            res.data.dropDownListData && res.data.dropDownListData[0].category
+    } else {
+      axios
+        .get(`/api/questions/${orgId}`)
+        .then(res => {
+          // make a conditon to check if editing or not(query param?!!)
+          const answers = {};
+          const edit = this.props.edit;
+          if (edit) {
+            // check the conditions ( 4 weeks + no helpful + no company response)
+            // send "seperate" request to get his answers (refactor)
+            const { getReviewAnswers: reviewDetails } = res.data;
+            reviewDetails[0].answers.map(answer => {
+              const {
+                answer: ans,
+                question: [question]
+              } = answer;
+              const number = question.number;
+              if (answers[number]) {
+                // think about this again;
+                answers[number] = ans;
+              } else {
+                answers[number] = ans;
+              }
+            });
+          }
+          const groupss = {};
+          res.data.groups.forEach(group => {
+            groupss[group._id] = {
+              title: group.group.text,
+              main: group.questions.filter(question => !question.isDependent),
+              dependant: group.questions.filter(
+                question => question.isDependent
+              )
+            };
+          });
+          this.setState({
+            groups: res.data,
+            groupss,
+            isLoading: false,
+            organization: res.data.organization,
+            email,
+            answers,
+            dropdownOptions:
+              res.data.dropDownListData && res.data.dropDownListData[0].category
+          });
+        })
+        .catch(err => {
+          // server error 500
+          const error =
+            err.response && err.response.data && err.response.data.error;
+          message.error(error || "Something went wrong");
         });
-      })
-      .catch(err => {
-        // server error 500
-        const error =
-          err.response && err.response.data && err.response.data.error;
-        message.error(error || "Something went wrong");
-      });
+    }
   }
 
-  // handleYesNo() => {
+  handleChange = e => {
+    const { answers } = this.state;
+    const { name, value } = e.target;
+    this.setState({
+      answers: { ...answers, [name]: value }
+    });
+  };
 
-  // }
+  handleCheckBox = () => {
+    const { hasAgreed } = this.state;
+    this.setState(
+      {
+        hasAgreed: !hasAgreed
+      },
+      () => {
+        this.runValidation2();
+      }
+    );
+  };
+
+  handleReviewChange = e => {
+    const { review } = this.state;
+    const { value } = e.target;
+    this.setState({
+      review: { ...review, overallReview: value }
+    });
+  };
+
+  handleSliderChange = (value, number) => {
+    const { answers } = this.state;
+    this.setState({
+      answers: { ...answers, [number]: value }
+    });
+  };
+
+  handleImageUpload = (value, number) => {
+    const { answers } = this.state;
+    this.setState({
+      answers: { ...answers, [number]: value }
+    });
+  };
+
+  handleRateChage = value => {
+    const { review } = this.state;
+    this.setState(
+      {
+        review: { ...review, rate: value }
+      },
+      () => {
+        this.runValidation2();
+      }
+    );
+  };
+
+  handleDateChage = (fromOrTo, value) => {
+    const { review } = this.state;
+    const { workPeriod } = review;
+    this.setState(
+      {
+        review: { ...review, workPeriod: { ...workPeriod, [fromOrTo]: value } }
+      },
+      () => {
+        this.runValidation2();
+      }
+    );
+  };
 
   showNextQestion = (groupId, next, other, set, num) => {
     const newGroups = { ...this.state.groupss };
@@ -158,10 +305,17 @@ class Review extends Component {
       // eslint-disable-next-line array-callback-return
       newDependant.map(question => {
         if (question.type === "number") {
-          set(`questions[${question.number}]`, null);
-        } else set(`questions[${question.number}]`, "");
+          this.setState({
+            answers: { ...this.state.answers, [question.number]: null }
+          });
+        } else {
+          this.setState({
+            answers: { ...this.state.answers, [question.number]: "" }
+          });
+        }
       });
     }
+
     while (typeof next !== "object" && next !== null) {
       // eslint-disable-next-line no-loop-func
       const nextQ = newDependant.find(question => question.number === next);
@@ -178,8 +332,14 @@ class Review extends Component {
       // eslint-disable-next-line array-callback-return
       newDependant.map(question => {
         if (question.type === "number") {
-          set(`questions[${question.number}]`, null);
-        } else set(`questions[${question.number}]`, "");
+          this.setState({
+            answers: { ...this.state.answers, [question.number]: null }
+          });
+        } else {
+          this.setState({
+            answers: { ...this.state.answers, [question.number]: "" }
+          });
+        }
       });
     }
     group.main = newMain.sort((a, b) => a.number - b.number);
@@ -188,60 +348,168 @@ class Review extends Component {
     this.setState({ groupss: newGroups });
   };
 
-  handleSubmit = (values, { setSubmitting }) => {
+  runValidation2 = () => {
+    const { organization } = this.state;
+    const values = {
+      answers: this.state.answers,
+      comments: this.state.comments,
+      review: this.state.review,
+      hasAgreed: this.state.hasAgreed
+    };
+    validationSchema[organization.category]
+      .validate(values, { abortEarly: false })
+      .catch(errors => {
+        const errs = {
+          answers: {},
+          review: {
+            workPeriod: {
+              from: "",
+              to: ""
+            },
+            rate: "",
+            overallReview: ""
+            // voiceReview: ""
+          },
+          hasAgreed: false
+        };
+        errors.inner.map(err => {
+          if (err.path.includes("answers")) {
+            const num = err.path.split(".")[1];
+            errs.answers[num] = err.message;
+          } else if (err.path.includes("workPeriod")) {
+            const key = err.path.split(".")[2];
+            errs.review.workPeriod[key] = err.message;
+          } else if (err.path.includes("rate")) {
+            errs.review.rate = err.message;
+          } else {
+            errs.hasAgreed = err.message;
+          }
+        });
+        this.setState({ errors: errs });
+      });
+  };
+
+  runValidation = () =>
+    new Promise((resolve, reject) => {
+      const { organization } = this.state;
+      const { user } = this.state;
+      const values = {
+        answers: this.state.answers,
+        comments: this.state.comments,
+        review: this.state.review,
+        hasAgreed: this.state.hasAgreed
+      };
+      validationSchema[organization.category]
+        .validate(values, { abortEarly: false })
+        .then(values => {
+          resolve(values);
+        })
+        .catch(errors => {
+          const errs = {
+            answers: {},
+            review: {
+              workPeriod: {
+                from: "",
+                to: ""
+              },
+              rate: "",
+              overallReview: ""
+              // voiceReview: ""
+            },
+            hasAgreed: false
+          };
+          errors.inner.map(err => {
+            if (err.path.includes("answers")) {
+              const num = err.path.split(".")[1];
+              errs.answers[num] = err.message;
+            } else if (err.path.includes("workPeriod")) {
+              const key = err.path.split(".")[2];
+              errs.review.workPeriod[key] = err.message;
+            } else if (err.path.includes("rate")) {
+              errs.review.rate = err.message;
+            } else {
+              errs.hasAgreed = err.message;
+            }
+          });
+          reject(errs);
+        });
+    });
+
+  handleSubmit = e => {
+    e.preventDefault();
+    this.setState({ isSubmitting: true });
     const { organization } = this.state;
     const { user } = this.state;
-    const review = {
-      values,
-      organization,
-      user
+    const values = {
+      answers: this.state.answers,
+      comments: this.state.comments,
+      review: this.state.review,
+      hasAgreed: this.state.hasAgreed
     };
-    axios
-      .post(API_POST_REVIEW_URL, review)
-      .then(res => {
-        this.props.history.push(THANKYOU_URL, {
-          orgType: organization.category,
-          orgId: res.data,
-          orgName: organization.name
-        });
+
+    this.runValidation(organization.category, values)
+      .then(vals => {
+        const review = {
+          values,
+          organization,
+          user
+        };
+        if (this.state.isEditing) {
+          // update the same review
+          axios
+            .put(`/api/review/${this.state.reviewId}`, review)
+            .then(res => {
+              this.setState({ isSubmitting: false });
+              this.props.history.push(THANKYOU_URL, {
+                orgType: organization.category,
+                orgId: res.data,
+                orgName: organization.name
+              });
+            })
+            .catch(err => {
+              const error =
+                err.response && err.response.data && err.response.data.error;
+              message.error(error || "Something went wrong");
+              // server error 500, maybe redirect to 500.error page??!!
+              this.setState({ isSubmitting: false });
+            });
+        } else {
+          axios
+            .post(API_POST_REVIEW_URL, review)
+            .then(res => {
+              this.setState({ isSubmitting: false });
+              this.props.history.push(THANKYOU_URL, {
+                orgType: organization.category,
+                orgId: res.data,
+                orgName: organization.name
+              });
+            })
+            .catch(err => {
+              const error =
+                err.response && err.response.data && err.response.data.error;
+              message.error(error || "Something went wrong");
+              // server error 500, maybe redirect to 500.error page??!!
+              this.setState({ isSubmitting: false });
+            });
+        }
       })
       .catch(err => {
-        const error =
-          err.response && err.response.data && err.response.data.error;
-        message.error(error || "Something went wrong");
-        // server error 500, maybe redirect to 500.error page??!!
-        setSubmitting(false);
+        this.setState({ errors: err });
       });
   };
 
   render() {
     const {
       groupss,
-
-      organization: { name, category }
+      organization: { name, category },
+      errors,
+      isSubmitting
     } = this.state;
-
     const { history } = this.props;
-
     const staticQuestion = STATIC_QUESTIONS(category);
 
     const { isLoading } = this.state;
     if (isLoading) return <Loading />;
-
-    const initialValues = {
-      questions: initQueestionsValues[this.state.organization.category],
-      comments: initQueestionsValues[this.state.organization.category],
-      review: {
-        workPeriod: {
-          from: "",
-          to: ""
-        },
-        rate: 0,
-        overallReview: ""
-        // voiceReview: ""
-      },
-      hasAgreed: false
-    };
 
     if (!this.state && !this.state.groups[0]) {
       return null;
@@ -274,131 +542,105 @@ class Review extends Component {
         </Header>
 
         <section className="review-body">
-          <Formik
-            initialValues={initialValues}
-            onSubmit={this.handleSubmit}
-            validationSchema={
-              validationSchema[this.state.organization.category]
-            }
-          >
-            {({
-              values,
-              isSubmitting,
-              handleChange,
-              errors,
-              setFieldValue
-            }) => {
-              return (
-                <FormWrapper>
-                  <Form>
-                    <Question
-                      {...values}
-                      question={staticQuestion[0]}
-                      setFieldValue={setFieldValue}
-                      category={this.state.organization.category}
-                    />
-                    <div>
-                      {Object.keys(groupss).map(groupId => {
-                        const group = groupss[groupId];
-                        if (group && group.title) {
+          <form onSubmit={this.handleSubmit}>
+            <FormWrapper>
+              <Question
+                question={staticQuestion[0]}
+                category={this.state.organization.category}
+                handleChange={this.handleDateChage}
+                state={this.state}
+                onBlur={this.runValidation2}
+              />
+              <div>
+                {Object.keys(groupss).map(groupId => {
+                  const group = groupss[groupId];
+                  if (group && group.title) {
+                    return (
+                      <div key={groupId}>
+                        <h2>{group.title}</h2>
+                        {group.main.map(question => {
                           return (
-                            <div key={groupId}>
-                              <h2>{group.title}</h2>
-                              {group.main.map(question => {
-                                return (
-                                  <Question
-                                    key={question._id}
-                                    showNextQestion={this.showNextQestion}
-                                    groupId={groupId}
-                                    values={values}
-                                    handleChagne={handleChange}
-                                    question={question}
-                                    errors={errors}
-                                    setFieldValue={setFieldValue}
-                                    dropdownOptions={this.state.dropdownOptions}
-                                    onChangeHandlers={this.handleChange}
-                                  />
-                                );
-                              })}
-                            </div>
+                            <Question
+                              key={question._id}
+                              question={question}
+                              showNextQestion={this.showNextQestion}
+                              groupId={groupId}
+                              dropdownOptions={this.state.dropdownOptions}
+                              handleChange={this.handleChange}
+                              state={this.state}
+                              handleSliderChange={this.handleSliderChange}
+                            />
                           );
-                        }
-                        return null;
-                      })}
-                    </div>
-                    <div className="questions">
-                      <Question
-                        {...values}
-                        handleChagne={handleChange}
-                        question={rateQ ? rateQ : ""}
-                        setFieldValue={setFieldValue}
-                        category={this.state.organization.category}
-                      />
-                      <Question
-                        {...values}
-                        handleChagne={handleChange}
-                        question={staticQuestion[1]}
-                        category={this.state.organization.category}
-                      />
-                      {/* The voice questions in the next sprint */}
-                      {/* <Question
-                        {...values}
-                        handleChagne={handleChange}
+                        })}
+                      </div>
+                    );
+                  }
+                  return null;
+                })}
+              </div>
+              <div className="questions">
+                <Question
+                  question={staticQuestion[2]}
+                  category={this.state.organization.category}
+                  handleChange={this.handleRateChage}
+                  state={this.state}
+                  runValidation={this.runValidation2}
+                />
+                <Question
+                  question={staticQuestion[1]}
+                  category={this.state.organization.category}
+                  handleChange={this.handleReviewChange}
+                  state={this.state}
+                />
+                {/* The voice questions in the next sprint */}
+                {/* <Question
                         question={staticQuestion[3]}
                         category={this.state.organization.category}
                       /> */}
-                    </div>
-                    <UserAgreement>
-                      <Level2Header>Submit your review</Level2Header>
-                      <CheckboxWrapper>
-                        <Field name={`hasAgreed`} id="agreement">
-                          {({ field, form }) => (
-                            <Checkbox
-                              {...field}
-                              id="agreement"
-                              style={{ marginTop: "4px" }}
-                            />
-                          )}
-                        </Field>
-
-                        <AgreementLabel htmlFor="agreement">
-                          I agree to the earwig{" "}
-                          <LinkSpan
-                            target="_blank"
-                            to={TERMS_OF_USE_URL}
-                            color={organizations[category].primary}
-                          >
-                            Terms of Use.
-                          </LinkSpan>{" "}
-                          This review of my experience with this current or
-                          former {category} is truthful.
-                        </AgreementLabel>
-                        <ErrorMessage name={`hasAgreed`}>
-                          {msg => (
-                            <StyledErrorMessage>{msg}</StyledErrorMessage>
-                          )}
-                        </ErrorMessage>
-                      </CheckboxWrapper>
-                    </UserAgreement>
-                    <SubmitButton
-                      type="submit"
-                      size="large"
-                      disabled={isSubmitting}
-                      orgType={category}
+              </div>
+              <UserAgreement>
+                <Level2Header>Submit your review</Level2Header>
+                <CheckboxWrapper>
+                  <Checkbox
+                    onChange={this.handleCheckBox}
+                    style={{ marginTop: "4px" }}
+                    checked={this.state.hasAgreed}
+                  >
+                    <AgreementLabel
+                      htmlFor="agreement"
+                      style={{ pointerEvents: "none" }}
                     >
-                      {isSubmitting && (
-                        <Spin
-                          indicator={antIcon}
-                          style={{ marginRight: ".5rem" }}
-                        />
-                      )}
-                      Submit your review
-                    </SubmitButton>
-                  </Form>
-                </FormWrapper>
-              );
-            }}
-          </Formik>
+                      I agree to the earwig{" "}
+                      <LinkSpan
+                        target="_blank"
+                        to={TERMS_OF_USE_URL}
+                        color={organizations[category].primary}
+                      >
+                        Terms of Use.
+                      </LinkSpan>{" "}
+                      This review of my experience with this current or former{" "}
+                      {category} is truthful.
+                    </AgreementLabel>
+                  </Checkbox>
+
+                  {errors && errors.hasAgreed && (
+                    <StyledErrorMessage>{errors.hasAgreed}</StyledErrorMessage>
+                  )}
+                </CheckboxWrapper>
+              </UserAgreement>
+              <SubmitButton
+                type="submit"
+                size="large"
+                disabled={isSubmitting}
+                orgType={category}
+              >
+                {isSubmitting && (
+                  <Spin indicator={antIcon} style={{ marginRight: ".5rem" }} />
+                )}
+                Submit your review
+              </SubmitButton>
+            </FormWrapper>
+          </form>
         </section>
       </ReviewWrapper>
     );
