+++ client/src/Components/index.js
@@ -25,7 +25,6 @@ import Thankyou from "./Pages/ThankYou";
 import EditProfile from "./Pages/EditProfile";
 import DeleteProfile from "./Pages/DeleteProfile";
 import UserProfile from "./Pages/UserProfile";
-import Review from "./Pages/Review";
 import Profile from "./Pages/Profile";
 import Admin from "./Pages/Admin";
 import Search from "./Pages/Search";
@@ -37,7 +36,7 @@ import PrivateRoute from "./Common/PrivateRoute";
 import Reply from "./Pages/Profile/Reply";
 import ResetPassword from "./Pages/ResetPassword";
 
-import Test from "./Pages/Review/test";
+import Review from "./Pages/Review";
 
 import {
   FAQ,
@@ -59,7 +58,6 @@ import {
   ADD_PROFILE_URL,
   ADD_PROFILE_START_REVIEW_URL,
   ADMIN,
-  REVIEW_URL,
   CONFIRM_EMAIL_URL,
   INTRO_URL,
   USER_PROFILE_URL,
@@ -73,15 +71,9 @@ export default function index(props) {
   return (
     <>
       <Switch>
-        <Route exact path="/organization/:orgId/review" component={Test} />
-        <Route exact path="/review/:reviewId/edit" component={Test} />
-        <PrivateRoute
-          exact
-          minimumLevel="LEVEL3"
-          path={REVIEW_URL}
-          {...props}
-          Component={Review}
-        />
+        <Route exact path="/organization/:orgId/review" component={Review} />
+        <Route exact path="/review/:reviewId/edit" component={Review} />
+
         <PrivateRoute
           exact
           minimumLevel="LEVEL3"
